+++ ase/calculators/eam.py
@@ -10,7 +10,7 @@
 import os
 import numpy as np
 from ase.test import NotAvailable
-from ase.calculators.neighborlist import NeighborList
+from ase.neighborlist import NeighborList
 from ase.calculators.calculator import Calculator, all_changes
 from scipy.interpolate import InterpolatedUnivariateSpline as spline