+++ ase/io/jsv.py
@@ -11,7 +11,7 @@ import re
 import numpy as np
 
 import ase
-from ase.lattice.spacegroup import Spacegroup, crystal
+from ase.spacegroup import Spacegroup, crystal
 from ase.geometry import cellpar_to_cell, cell_to_cellpar