+++ doc/tutorials/neb/idpp2.py
@@ -1,4 +1,4 @@
-from ase.structure import molecule
+from ase.build import molecule
 from ase.neb import NEB
 from ase.calculators.emt import EMT
 from ase.optimize.fire import FIRE as QuasiNewton