+++ doc/tutorials/spacegroup/spacegroup-rutile.py
@@ -1,4 +1,4 @@
-from ase.lattice.spacegroup import crystal
+from ase.spacegroup import crystal
 
 a = 4.6
 c = 2.95