+++ doc/tutorials/spacegroup/spacegroup-skutterudite.py
@@ -1,4 +1,4 @@
-from ase.lattice.spacegroup import crystal
+from ase.spacegroup import crystal
 
 a = 9.04
 skutterudite = crystal(('Co', 'Sb'),